Remove typescript support in preparation for ejection
