+++ package.json
@@ -8,11 +8,6 @@
     "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/react-fontawesome": "^0.1.4",
     "@giantmachines/redux-websocket": "^1.1.2",
-    "@types/jest": "^24.0.15",
-    "@types/lodash": "^4.14.134",
-    "@types/node": "^12.0.10",
-    "@types/react": "^16.8.22",
-    "@types/react-dom": "^16.8.4",
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "classnames": "^2.2.6",
@@ -28,8 +23,7 @@
     "react-scripts": "3.0.1",
     "redux": "^4.0.1",
     "redux-saga": "^1.0.3",
-    "twemoji": "^12.0.4",
-    "typescript": "^3.5.2"
+    "twemoji": "^12.0.4"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
